requirements.txt
flask
flask_bootstrap
flask_sqlalchemy 
flask_wtf 
flask_login 
flask_mail
wtforms 
pyusb
paho-mqtt
pyyaml